+++ packages/event-log/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @claudexor/event-log
 
+## 3.12.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @claudexor/schema@3.12.5
+  - @claudexor/util@3.12.5
+
 ## 3.12.4
 
 ### Patch Changes